Fix sdist and relevant CI, update libsonata. (#108)
Adds a MANIFEST.in that should include most needed files, and a CI that always attempts to build sdist and wheels. Upload only when tagging the repository.

Also updates PyBind11 to a version that can handle Python 3.11 and uses a more modern HDF5 with less bugs when building wheels.

NB: The source distribution is still too large! Test file size should be reduced, maybe the dependencies built in the CI (libsonata, highfive)